deep_learning_project/requirements.txt
fastapi~=0.111.1
MyApplication~=0.1.0
pydantic~=2.8.2
uvicorn~=0.30.3
python-dotenv~=1.0.1
numpy~=1.26.1
joblib~=1.3.2
pandas~=2.1.1
scipy~=1.11.4
scikit-learn~=1.3.2
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
